moved cmd version and exportregistries
