+++ internal/cli/cmd/exportregistry.go
@@ -1,0 +1,37 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"github.com/epinio/epinio/internal/cli/usercmd"
+	"github.com/pkg/errors"
+	"github.com/spf13/cobra"
+)
+
+// NewExportRegistriesCmd returns a new 'epinio info' command
+func NewExportRegistriesCmd(client *usercmd.EpinioClient) *cobra.Command {
+	return &cobra.Command{
+		Use:   "export-registries",
+		Short: "List export registries",
+		Long:  "List export registries",
+		Args:  cobra.NoArgs,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			cmd.SilenceUsage = true
+
+			err := client.ExportregistryList(cmd.Context())
+			if err != nil {
+				return errors.Wrap(err, "error listing export registries")
+			}
+			return nil
+		},
+	}
+}
